Create README.md
Python script to check the server directory. The device processes logs to provide information such as the number of requests per IP address, most visited content, and to detect suspicious activity (such as malicious attempts). The output is displayed in the terminal and stored in a CSV file for easy publishing.